Added Script, Fixed some level stuff
Added a character controller that allows you to move forwards, backwards, left, right, and jump. I am waiting on learning how to aim and shoot from the professor so the camera can be moved allowing you to turn.

I fixed some parts of the level that didn't let you jump, but more things will be added to not make the level look super boring.