Created new scene and stuff
made a first level scene which is where you will begin playing in the level. It is not complete (at all). Still need to add any obstacles or anything to make the level more interesting in the Ground section.

 The cannotPass is supposed to be where the player cannot move onto. I'll have to add some invisible walls or make something so the player doesn't venture off. That section is primarily reserved for any decorations to make it look nice. The ant hill will have to be made in blender and is not something the player will be able to go inside. I would add platforms, but i wouldn't want the player to stay on there and just shoot the enemies. If the enemies are able to follow by also getting on the platforms, then many types of obstacles can be added.

messed with the prefab a little, but still works with sewer level scene.

I added a skybox, but it's not an official one, i just needed something to be on the background so not everything would be dull and boring. it looks kinda funny LOL