Added a new scene, and some terrain
I made a new scene for the inside of the ant hill, which will be a boss kind of level so once you reached a certain number of rounds in the first level, you will go to the inside to fight the boss.

I added some terrain in both scenes, but will need to finish with the terrain in the firstlevel_scene.