Finished levels with music
All levels are technically finished.
All that would be needed is adding a little bit more things like obstacles and adding more hazards, like the lava and the toxic.

The music that goes with the scene is there. If you need to adjust the volume, you are free to do so.

You can use the levels First, Second, and Third ones for final testing.