Fixed the EnemySpawner and added OffMesh links
Fixed the enemy spawner so that a max amount of enemies will show up.
The check to move to the next level is based on how many enemies are left in the level. (We should probably add a enemies remaining counter)

Offmesh links have been added on the levels that they are needed on so the enemies can get down faster and are able to jump to different obstacles.