Fixed some stuff
Fixed the controller for the player
   although it moves kind of odd, it is still pretty good.

Added changes to scenes, like fixing the enemies.

All that would be left is adding and changing the sprite depending on what gun you are on.
An end scene.
An enemies remaining text on the canvas to let the player know how many enemies are left.
And maybe instead of a spawner, selecting which powerups show up where and how many there should be.